Also decorate OperatorBinding in LibraryExpressionBuilder
