+++ Cql/Cql.Compiler/ExpressionBuilderContext.Debugging.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Diagnostics;
-using System.Linq.Expressions;
-using Hl7.Cql.Abstractions;
-using Hl7.Cql.Abstractions.Exceptions;
+﻿using System.Diagnostics;
 using elm = Hl7.Cql.Elm;
 
 namespace Hl7.Cql.Compiler;
@@ -17,37 +13,6 @@ partial class ExpressionBuilderContext : IBuilderContext
     IBuilderContext? IBuilderContext.OuterContext => (IBuilderContext?)_outerContext ?? LibraryContext;
 
     BuilderContextInfo IBuilderContext.ContextInfo => BuilderContextInfo.FromElement(_element);
-
-    private sealed class OperatorBindingRethrowDecorator : OperatorBinding
-    {
-        private readonly ExpressionBuilderContext _owningExpressionBuilderContext;
-
-        public OperatorBindingRethrowDecorator(
-            ExpressionBuilderContext owningExpressionBuilderContext,
-            OperatorBinding inner)
-        {
-            _owningExpressionBuilderContext = owningExpressionBuilderContext;
-            Inner = inner;
-        }
-
-        public OperatorBinding Inner { get; }
-
-        public override Expression Bind(CqlOperator @operator, Expression runtimeContext, params Expression[] parameters)
-        {
-            try
-            {
-                return Inner.Bind(@operator, runtimeContext, parameters);
-            }
-            catch (CqlException<ExpressionBuildingError>)
-            {
-                throw;
-            }
-            catch (Exception e)
-            {
-                throw _owningExpressionBuilderContext.NewExpressionBuildingException(e.Message, e);
-            }
-        }
-    }
 
     public string FormatMessage(string message, elm.Element? element = null)
     {
